refactor(helloworld): switch to InterationHandler correctly
